feat: implement AWS S3 as storage provider
These changes allows us to configure an AWS S3 bucket to be the destination
for images taken, just like any other storage provider.

The added configuration examples tells what's neccessarry to get this
going. Other than that, having signed in with AWS CLI tools on the machine
it's supposed to run on is required for now.